+++ src/main/java/roomescape/auth/dto/LoginResponse.java
@@ -1,4 +1,4 @@
-package roomescape.member.dto;
+package roomescape.auth.dto;
 
 public record LoginResponse(String name) {
 }
